form_toolkit/requirements.txt
flask>=3.0.0
reportlab>=4.0.0
pypdf>=4.0.0
llama-cloud>=1.0.0
httpx
anthropic>=0.40.0
python-dotenv>=1.0.0
